chore: rename ai package
