+++ packages/@react-spectrum/ai/README.md
@@ -1,3 +1,3 @@
-# @react-spectrum/s2-ai
+# @react-spectrum/ai
 
 This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
