+++ packages/@react-spectrum/ai/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@react-spectrum/s2-ai",
+  "name": "@react-spectrum/ai",
   "version": "1.0.0-alpha.1",
   "description": "Spectrum 2 UI AI components in React",
   "license": "Apache-2.0",
