+++ yarn.config.cjs
@@ -27,7 +27,7 @@ function enforceConsistentDependenciesAcrossTheProject({Yarn}) {
         } else if (
           dependency.workspace.ident === '@react-spectrum/s2' ||
           dependency.workspace.ident === '@react-spectrum/s2-icon-builder' ||
-          dependency.workspace.ident === '@react-spectrum/s2-ai'
+          dependency.workspace.ident === '@react-spectrum/ai'
         ) {
           dependency.update('^19.0.0-rc.1');
         } else if (dependency.workspace.ident === '@react-spectrum/codemods') {
@@ -58,7 +58,7 @@ function enforceConsistentDependenciesAcrossTheProject({Yarn}) {
       workspace.ident !== '@internationalized/string-compiler' &&
       workspace.ident !== 'tailwindcss-react-aria-components' &&
       workspace.ident !== '@react-spectrum/s2' &&
-      workspace.ident !== '@react-spectrum/s2-ai' &&
+      workspace.ident !== '@react-spectrum/ai' &&
       workspace.manifest.rsp?.type !== 'cli'
     ) {
       workspace.set('dependencies.@swc/helpers', '^0.5.0');
